only make img that doesn't exist

+++ index.js
@@ -17,15 +17,17 @@ function reduceImg( dimensions, expectedWidth ) {
 fs.readdirSync( opts.dirs.original ).forEach( file => {
     const dimensions = sizeOf( path.join( opts.dirs.original, file ) )
     Object.keys(opts.resos).forEach( reso => {
-        reso = opts.resos[reso]
-        // create dir if necessary
-        let dir = path.join( opts.dirs.imgFolder, reso.prefix )
-        if ( !fs.existsSync( dir ) ) fs.mkdirSync(dir) 
-        // get new dimensions
-        const newDims = reduceImg( dimensions, reso.width )
-        // create new resized img
-        resizeImg( fs.readFileSync( path.join( opts.dirs.original, file ) ), {width: newDims.width, height: newDims.height} ).then( buf => {
-            fs.writeFileSync( path.join( opts.dirs.imgFolder, reso.prefix, file ), buf)
-        })
+        if ( !fs.existsSync( path.join( opts.dirs.imgFolder, opts.resos[reso].prefix, file ) ) ) {
+            reso = opts.resos[reso]
+            // create dir if necessary
+            let dir = path.join( opts.dirs.imgFolder, reso.prefix )
+            if ( !fs.existsSync( dir ) ) fs.mkdirSync(dir) 
+            // get new dimensions
+            const newDims = reduceImg( dimensions, reso.width )
+            // create new resized img
+            resizeImg( fs.readFileSync( path.join( opts.dirs.original, file ) ), {width: newDims.width, height: newDims.height} ).then( buf => {
+                fs.writeFileSync( path.join( opts.dirs.imgFolder, reso.prefix, file ), buf)
+            })
+        }
     })
 })
